backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2
